Fix: Remove package constants
